Add download button for json data
