+++ client/src/components/Template/index.ts
@@ -1,4 +1,5 @@
 import Template from './Template';
 import './styles.scss';
 
+export * from './Template';
 export default Template;
