+++ client/src/components/TemplateHelp/index.ts
@@ -1,3 +1,5 @@
 import TemplateHelp from './TemplateHelp';
 
+import './styles.scss';
+
 export default TemplateHelp;
