+++ client/src/components/TemplateHelp/styles.scss
@@ -3,6 +3,10 @@
     margin: 1rem;
   }
 
+  .button-container {
+    margin-bottom: 1rem;
+  }
+
   .helper-description {
     margin-bottom: 1rem;
   }
